move config definition outside common
